+++ src/basic/components/CartList.tsx
@@ -1,5 +1,7 @@
 import { calculateItemTotal } from '../models/cart';
 import { CartItem, Coupon } from '../types';
+import { CartCouponList } from './cart/CartCouponList';
+import { PaymentInfo } from './cart/PaymentInfo';
 
 interface CartPageProps {
   cart: CartItem[];
@@ -129,69 +131,12 @@ export const CartPage = ({
 
       {cart.length > 0 && (
         <>
-          <section className='bg-white rounded-lg border border-gray-200 p-4'>
-            <div className='flex items-center justify-between mb-3'>
-              <h3 className='text-sm font-semibold text-gray-700'>쿠폰 할인</h3>
-              <button className='text-xs text-blue-600 hover:underline'>쿠폰 등록</button>
-            </div>
-            {coupons.length > 0 && (
-              <select
-                className='w-full text-sm border border-gray-300 rounded px-3 py-2 focus:outline-none focus:border-blue-500'
-                value={selectedCoupon?.code || ''}
-                onChange={(e) => {
-                  const coupon = coupons.find((c) => c.code === e.target.value);
-                  if (coupon) applyCoupon(coupon);
-                  else applyCoupon(null);
-                }}
-              >
-                <option value=''>쿠폰 선택</option>
-                {coupons.map((coupon) => (
-                  <option key={coupon.code} value={coupon.code}>
-                    {coupon.name} (
-                    {coupon.discountType === 'amount'
-                      ? `${coupon.discountValue.toLocaleString()}원`
-                      : `${coupon.discountValue}%`}
-                    )
-                  </option>
-                ))}
-              </select>
-            )}
-          </section>
-
-          <section className='bg-white rounded-lg border border-gray-200 p-4'>
-            <h3 className='text-lg font-semibold mb-4'>결제 정보</h3>
-            <div className='space-y-2 text-sm'>
-              <div className='flex justify-between'>
-                <span className='text-gray-600'>상품 금액</span>
-                <span className='font-medium'>{totals.totalBeforeDiscount.toLocaleString()}원</span>
-              </div>
-              {totals.totalBeforeDiscount - totals.totalAfterDiscount > 0 && (
-                <div className='flex justify-between text-red-500'>
-                  <span>할인 금액</span>
-                  <span>
-                    -{(totals.totalBeforeDiscount - totals.totalAfterDiscount).toLocaleString()}원
-                  </span>
-                </div>
-              )}
-              <div className='flex justify-between py-2 border-t border-gray-200'>
-                <span className='font-semibold'>결제 예정 금액</span>
-                <span className='font-bold text-lg text-gray-900'>
-                  {totals.totalAfterDiscount.toLocaleString()}원
-                </span>
-              </div>
-            </div>
-
-            <button
-              onClick={completeOrder}
-              className='w-full mt-4 py-3 bg-yellow-400 text-gray-900 rounded-md font-medium hover:bg-yellow-500 transition-colors'
-            >
-              {totals.totalAfterDiscount.toLocaleString()}원 결제하기
-            </button>
-
-            <div className='mt-3 text-xs text-gray-500 text-center'>
-              <p>* 실제 결제는 이루어지지 않습니다</p>
-            </div>
-          </section>
+          <CartCouponList
+            selectedCoupon={selectedCoupon}
+            coupons={coupons}
+            applyCoupon={applyCoupon}
+          />
+          <PaymentInfo totals={totals} completeOrder={completeOrder} />
         </>
       )}
     </div>
